requirements.txt
fastapi==0.115.0
httpx==0.27.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
slack_sdk==3.33.1
uvicorn==0.30.6
starlette==0.38.6
tenacity==9.0.0
